Clean up a bunch of leftovers.
There was quite a bit of commentary and commented-out code in the .h
file, left over from when this file was created out of the one for fifo
tasking.  None of it is apropos any more.  Remove it.  In a few cases,
add back little header comments.

While here, I removed one more reference to QTHREAD_MULTINODE, which is
multinode support we in Chapel don't need or want.