Note where pre-resolution passes should handle RET_CONST_REF
In a few cases it seems clear that they should handle RET_REF and
RET_CONST_REF in the same way, so I adjusted those. For others, I added a
TODO comment indicating where the code should be adapted to handle
RET_CONST_REF once we support it from parsing.