Add an explicit cast to quiet a Coverity complaint.
Coverity complained about returning the 'overflowed or truncated' value
of hwpar.  To quiet this, explicitly cast the unsigned long int value of
chpl_qt_getenv_num() to the int32_t type of hwpar when assigning the
former to the latter.  And to make the resulting code look less funny,
do the same for the num_wps and num_sheps assignments that follow, even
though Coverity didn't mention either of those.

(The Coverity complaint was new because the Qthreads tasking shim is new
in the runtime.  We don't have Coverity scan the third-party packages,
where this code used to reside.)