Applies excepted identifiers to later uses in same chain
Before, if a module M used a module A with an except on 'foo', any modules A
used which defined 'foo' would still act as a match.  Now, A's uses are treated
slightly different when A is used in a limited fashion.  This does not affect
the uses when trying to resolve A's symbols.

Implementation notes: I chose to do this by creating copies of the original use
and adding all the previous use's excepted identifiers to their except list.
These uses would often be created during processImportExprs, which traverses
the list of all UseExprs, so that part of scopeResolve needed to adjust for
these uses that should never be placed in the tree (even in the special use
list for each module and its block).  Luckily, these new uses could be
distinguished from the other uses by virtue of not having a parentExpr (even
though all other uses are removed from the tree, that action happens as part of
processImportExprs, so we would never run into them before updating their
mod field), so we could determine whether to skip trying to add the new use to a
block's use list based on that absence.