Adjust chpl_anycomplex.re to always return ref
Before this change, that function returned the result of creal() (e.g.)
as a value if !setter.

But, in making ref-return-intent functions return by const ref, this
function can no longer return a ref to the function local/stack allocated
result of creal(). So, I adjust it to always use complex_get_real, which
returns a ref.

David tells me that complex_get_real does not seem to show reduced
performance when compared to creal() in benchmarks.