Add future for chpldoc Sphinx bug Kyle found
Kyle discovered that if a top level module symbol is defined after a submodule,
our Sphinx interaction doesn't currently have a way to know that we meant it to
be associated with a different module than the one most recently defined.  Thus:

module Foo {
  module Bar {
     proc baz () {...}
  }

  proc gob () {...}
}

associates gob with Bar instead of Foo, as intended.  There are several
potential solutions to this problem:

1) we could always output submodules last, though that may cause sibling
submodules to be denoted as in a parent/child relationship (I'm not sure if
this will actually happen, though).

2) we could have Sphinx look for a marker to denote the end of a module.  I'm
not sure if there's a way to do that, but it seems at least possible

3) we could skip this problem entirely by having every module be in its own
file, something that Thomas had advocated before but neither of us got around
to doing.  Funneling the output might be a little tricky, but it seems doable.

I'd be tempted to try option 3 first, just because I'm more familiar with that
code (and option 1 seems like a hack).  I'm adding a future, though, in case
not getting to this right away means that it gets forgotten.