Consider returning from a RET_CONST_REF function as rvalue use
In cullOverReferences, the function refNecessary determines whether a
given variable (that is storing the result of a call to a
ref-return-intent-function) is used as an lvalue or is only used in an
rvalue context.

This patch adjusts refNecessary to consider returning the variable from a
function with RET_CONST_REF return tag to be a rvalue use, where before
all returns were considered lvalue uses.

This change was motivated by - and fixes - this test:
  test/arrays/deitz/myarray/test_array_class_trial5b.chpl