Adjust preFold to allow returning constant variables by const ref
As demonstrated in this test:

  test/functions/dinan/const_getter.chpl

a function with RET_CONST_REF ret tag function should be able to return a
reference to a global constant. This patch adjusts checks in preFold to
allow that to happen.

Some of the changes here are workarounds - it would be better to fully
integrate the idea of 'const ref' into function resolution (e.g. to flow
a 'const ref' type from where it is generated to where it is used).