Add initial support for CHPL_MEM=jemalloc
Overall
-------

This adds support for Jason Evan's memory allocator (jemalloc), which
"emphasizes fragmentation avoidance and scalable concurrency support." Note
that this doesn't actually add the jemalloc source code yet, just the
supporting infrastructure. The source will be added in a separate commit. The
motivation for using jemalloc is the same as tcmalloc: we want an allocator
that performs well in multi-threaded situations and most cstdlib versions
don't.  Unfortunately we ran into trouble trying to make tcmalloc our default
and it didn't build in a lot of configurations. jemalloc should be much simpler
to get working since it's C instead of C++. Additionally jemalloc is a much
more active project both in terms of development and use. Some of the major
users are: FreeBSD, Firefox, Varnish (memcache), Facebook, Intel (memkind),
Rust, HPX-5 (not big, but spiritually related to us) and a bunch of other large
projects/companies (https://github.com/jemalloc/jemalloc/wiki/Adoption)

To enable jemalloc I basically copied code used to support our other allocators
and modified it for use with jemalloc. The runtime shim is almost identical to
cstdlib and the third-party makefiles and chplenv stuff are pretty similar to
tcmalloc.

We're not doing anything too fancy when we build jemalloc. We're using a
vanilla version of jemalloc 4.0.4 with the default config (no real options
passed to configure) and then we just build/install the static only versions.

Portability
-----------

Note that this patch does _not_ add support multi-locale configs that require a
heap (gasnet segment large/fast and ugni), but this should allow us to
experiment and get a feel for performance compared to cstdlib and tcmalloc.

I did a full standard and gasnet testing run as well as a `make check` on
  - darwin
  - linux64
  - cray-xc
    - PrgEnv-gnu
    - PrgEnv-intel
    - PrgEnv-pgi

Currently, jemalloc fails to build with PrgEnv-cray, but I think we can work
around those issues.

jemalloc 4 added support for cywgin, but there appear to be a couple of issues
with this. I ran into the same ones described in:

 - https://github.com/jemalloc/jemalloc/issues/285
 - https://cygwin.com/ml/cygwin/2015-10/msg00082.html

Performance
-----------

For a few spot checks, the performance seems to be as good if not better than
tcmalloc! Sample of perf for Simple MiniMD (seconds):

|           | cstdlib | tcmalloc | jemalloc |
| --------- | ------- | -------- | -------- |
|--local    | 2.20    | 0.38     | 0.38     |
|--no-local | 2.50    | 0.90     | 0.70     |

TODOs
-----

 - Add support for multi-locale configs that require a heap (gasnet segment
   large/fast and ugni.) A possibly lead towards this might be:
   http://jemalloc-discuss.canonware.narkive.com/FzSQ4Qv4/need-help-in-porting-jemalloc
   I think we basically have to create a custom chunk allocator.

 - See if there's a way to implement chpl_goodAllocSize() smartly: possibly
   nallocx(). I think this would require using the extended API. Note that the
   extended API has undefined behavior for allocations of size 0 so an extra
   check would have to be added to our memory shim.

   If we use the extended API we may want to consider updating our API to
   include a function for sized deallocation that uses `sdallocx()`.
   Apparently this can have some significant performance benefits:

    - https://isocpp.org/files/papers/n3778.html