Add two more tests
These verify the behavior of the except keyword when multiple uses in a chain
utilize it.  One ensures that symbols excluded in the inner except are still
not visible when the outer module is used with a different except.  The other
ensures that multiple excepts don't cause unnecessary loss of identifiers in
the inner module use.