Add two tests and a future
One test ensures that multiple definitions of a function are excluded when the
name is specified, rather than just the first one that would be encountered.
The other test ensures that we can specify classes in the except list and the
response will be valid.  The future exposes a slightly problem with the latter,
which is that though we may not allow any direct matches to the type name, we
will still allow functions that return that type name without comment.