Replace var numCores with proc numPUs().
The locale.numCores variable was misnamed and limited: it actually
contained the number of hardware threads, not cores, that the OS had
made available.  There was no way to find out the number of cores the
OS had made available, nor any way to find out the actual (not
OS-limited) number of threads or hardware cores.  Here we replace
this variable with proc numPUs(), which can answers all those
questions.

Note that in tests where numPUs(physical=true) replaces numCores,
this means that now we're using the number of cores where before we
were using the number of hardware threads.  I'll ask for guidance on
this part of the change when it is reviewed.

For reference, the tests whose behavior actually depends in some way
on this value are just these 4:
  exercises/Mandelbrot/solutions/mandelbrot4-taskpar.chpl
  exercises/Mandelbrot/solutions/mandelbrot5-taskpar-dist.chpl
  users/aroonsharma/CyclicZipOpt.chpl
  users/aroonsharma/MyBlockCyclic.chpl

These other tests just printed the value; they didn't use it
computationally:
  functions/jturner/first-class-generic-function.chpl
  modules/standard/machine/numPUs.chpl (was numCores.chpl)
  multilocale/bradc/locales/queryLocaleStats.chpl
  release/examples/primers/locales.chpl
  studies/cholesky/jglewis/version2/elemental/block_distribution/test_elemental_cholesky_block.chpl
  studies/cholesky/jglewis/version2/elemental/fully_blocked/test_fully_blocked_elemental_cholesky.chpl
  studies/cholesky/jglewis/version2/elemental/strided/test_elemental_cholesky_explicitly_symmetric_strided.chpl
  studies/cholesky/jglewis/version2/elemental/test_elemental_cholesky.chpl
  studies/cholesky/jglewis/version2/elemental/unsymmetric_indices/test_elemental_cholesky_unsymmetric_ranges.chpl
  studies/graph500/kristyn/Graph500_1D_onV/main.chpl
  studies/graph500/v2/main.chpl
  studies/graph500/v3/main.chpl
  studies/ssca2/kristyn/SSCA2_main.chpl
  users/jglewis/test_cyclic_dist.chpl
  users/vass/internal-error-message.chpl

And finally, these tests only mentioned numCores in a comment:
  test/functions/iterators/angeles/checkAdvancedIters/checkAdaptiveWSIter.chpl
  test/functions/iterators/angeles/checkAdvancedIters/checkDynamicIter.chpl
  test/functions/iterators/angeles/checkAdvancedIters/checkGuidedIter.chpl
  test/functions/iterators/angeles/distAdaptativeWS.chpl
  test/functions/iterators/angeles/distAdaptativeWSv1.chpl
  test/functions/iterators/angeles/distAdaptativeWSv2.chpl
  test/functions/iterators/angeles/dynamic.chpl
  test/functions/iterators/angeles/guided.chpl
  gpu/sidelnik/hpcc/stream32-domain.chpl
  gpu/sidelnik/hpcc/stream32-zipper.chpl
  gpu/sidelnik/hpcc/stream64-domain.chpl
  gpu/sidelnik/hpcc/stream64-zipper.chpl
  users/vass/barrierWF.test-1.chpl
  users/wmikanik/test1.chpl