Adjust buildValueFunction to choose RET_VALUE or RET_CONST_REF
It chooses RET_CONST_REF for ref-intent functions returning types that are not
POD.

In order to make this choice based on the return type, the return type must be
known, which means that the function must be resolved. Since at this point in
function resolution, we are part-way through resolving the 'setter' version of
the function, we cannot rely on the return type being already resolved.

In order to work around this issue in the order of resolution, resolve first
the RET_VALUE version of the function. If the return type ends up being not
POD, create and use a RET_CONST_REF version of the function.