Adjust refNecessary to fix bug
The test
 arrays/bradc/reindex/reindex
was failing apparently because of a problem in iterator -> array
promotion. The previous version of this code caused the bug, but
adjusting it to check for RET_CONST_REF instead of RET_REF resolves
the issue. It must have to do with how iterator functions interact
with refNecessary.