Clean up Spawn error handling
Some common (but possibly bone-headed) code patterns
were causing core dumps. For example

  var sub = spawn([...], stdout=PIPE);
  var count = 0;
  var line:string;
  while sub.stdout.readline(line) {
    count += 1;
  }

will core dump if the spawn command failed. The current error strategy for the
spawn command is to store whether or not the spawn suceeded as error codes in
the returned record.

This patch modifies several of the spawn methods to make sure that they are
working with a sucessfully spawned subprocess, so that the program above halts
with an error on the 'sub.stdout' if there was error spawning.