Allow better clean up of channels in a subprocess
The current usage pattern for a subprocess might involve buffering up
output during a subprocess.communicate() call. So it does not make sense
to close the channels in the subprocess.communicate() call itself.  The
channels should be automatically closed when the subprocess record goes
out of scope. However, it is useful to also have a more explicit way of
closing the channels.

Therefore, this commit adds subprocess.close() methods and updates the
tests to call subprocess.close().

It also updates the Spawn documentation to indicate what will happen when
a subprocess record goes out of scope.

Also updated the various places in Spawn.chpl where the stdin
channel is closed so that it is safe to call more than one of these.
(the problem was that .commit could be called on an already closed
 channel).