Error messages and skip of symbol matching for names not in 'only' list
Will generate an error message if the identifier specified in an 'only' list is
not a public symbol, or is not found in the scope being searched.  Also prevents
looking in the current use if the name looked for is not in the 'only' list.
This does not yet dive into the modules used by the one specified with an
'only', that implementation still had problems.

Not that this commit doesn't quite work yet, there seems to be an unrelated
double free/corruption occuring.