Run these tests only on Mac OS X.
How errors in spawning are reported differs between Mac OS X and Linux. Mac OS
X creates an error if the file does not exist right away, but on Linux such an
error is only detectable by the exit code of the subprocess. We already have a
test checking the exit code of the subprocess.