Enable ldflags to stack like ccflags
If you specify multiple --ccflags options, e.g.
  --ccflags -O1 --ccflags -O3
the C compiler will get the combination of the flags, e.g.
  -O1 -O3

This feature is useful when some compiler arguments come from
a package (or a script giving arguments needed to use a library e.g.).
This pattern happens a lot with 3rd party software. llvm-config
is an example that comes to mind. Since the arguments stack, a
user of a library can also add flags, e.g.

  chpl `library-get-compflags.sh` --ccflags -fPIC

will append to - but not replace - any ccflags from the script.

Generally speaking, we expect that a C compiler will handle several
options for the same thing, where the last takes precedence, e.g.
 -O1 -O3
is really the same as
 -O3

This PR updates --ldflags to receive the same treatment, and cleans up
--ccflags and --ldflags support so that they use a std::string instead of
a fixed length C string. It adds a test to verify this behavior of
ldflags.