Expanded buildBreadthFirstModuleList's logic to propogate 'only's and handle mix
'only' lists will now be propogated to their nested uses, and 'except'+'only'
interminglings will result in the proper list for the use in that chain.  My
initial implementation had resulted in an infinite loop, because I wasn't noting
that I'd already seen modules with 'only' properly.  I have expanded the notion
of what was already seen to take into account the multiple ways a module could
be used and properly skip traversing a use if it was seen in a more general
manner.

For instance, if you were looking at:

use F except A;

but had already seen this use of F:

use F except A, B;

then you would still want to traverse into the new use.  But if you had seen the
first already and were just encountering the second for the first time, you
wouldn't need to traverse the new use, as it would provide no new symbols.

Similarly, if you were looking at:

use F only A;

but had already seen this use of F:

use F only A, B;

then you would not want to traverse into the new use, because the second use
provided all the symbols available in it.  'Except' uses are more general than
'only' uses, unless the 'except' list contains any of the symbols in the 'only'
list.

This implementation probably needs to be cleaned up, but I was just so happy it
was working, I had to commit it!