Make functionResolution implement 'only' and update 'except' implementation
My functionResolution implementation for 'except' also didn't make use of the
std::find function when it could have, so I fixed that.  Performed a similar
check in that area to determine whether to dive into a use with an 'only' list.
With that change, and a .good file correction, all the tests I have written
for the 'only' keyword pass.

TODO:
- write a few more tests
- clean up the code somewhat
- commit the updated parser files
- add the new keywords to the various text modes that we maintain.