Add remaining tests
These tests verify the behavior of nested 'only' and 'except' use chains,
ensuring that we don't get stuck in an infinite compiler loop and that symbols
aren't leaked when they shouldn't be.  There is also a new test to ensure that
symbols which would be ignored because of the 'only' list also don't interfere
with finding ones of the same name in an outer scope.