Reverse numPUs() physical/logical arg name, improve chpldocing, etc.
This all reflects feedback during review.

Rename numPUs() 'physical' argument as 'logical' and reverse the default
value so that the semantic result is the same.  There seemed to be too
many ways to misconstrue the meaning of 'physical'.

Improve the numPUs() chpldoc comment.  Add RsT entries for args and
return value, reduce duplication, and add a final paragraph that talks
about what things can cause the set of accessible PUs to be limited.

Remove explicit 'physical=true' actual args in calls everywhere.  These
just match the default.  But if the default is well chosen and well
documented, we shouldn't have to pass it explicitly anyway.

Make a pass over the tests, cleaning up.  Beyond removing explicit
'physical=true' actuals, change here.numPUs() to here.maxTaskPar in both
mandelbrot4-taskpar.chpl and mandelbrot5-taskpar-dist.chpl.  Based on
looking at the code, maxTaskPar is probably what is really wanted: an
estimate of the best concurrency width.  I also checked with Michael on
this change, and he okayed it verbally.