For comm=ugni hugepages optional: rm build-time warning, update docs.
We made hugepages optional with CHPL_COMM=ugni late last fall.  At the
time, I intended to remove the warning we had been issuing when the
compiler was run with CHPL_COMM=ugni and no craype-hugepages module
loaded, but I forgot to do so.  Remove that now.

Also, update the Cray platform documentation to reflect the now-optional
nature of hugepages.