Prevent casts from string to string.
I was observing that code like this:

    var s = "hello";
    proc foo(ref s)
    {
      var y = s:string;
      writeln(y);
    }
    foo(s);

generates a cast within foo(). It turns out that the code in
dropUnnecessaryCast did not take into account references, so that a cast
from ref(string) to string was not removed.

Updated that function to remove casts from ref(T) to T.

Updated StringCasts.chpl to no longer include a cast from
sttring to string.