Improve ugni Spawn halt to only fail with hugepages
Greg's recent work enables CHPL_COMM=ugni to work without a hugepages module
loaded. In that configuration, Spawn seems to work correctly.

Therefore, this commit updates the halt in Spawn.chpl that checks for
CHPL_COMM=ugni to also check for a huge pages module being loaded. This check
is specific to a Cray, but so is CHPL_COMM=ugni.