Add tests of the type behavior in an 'only' list
One exposes the behavior when a type is not provided on the 'only' list, but
a function that returns a variable of that type is provided and used.  Our hope
is that we can warning/error in this case in the future.  The other two tests
expose the behavior when a type is provided on the 'only' list.  Right now,
it seems as if the constructor is blocked, which doesn't make sense since it
should share the same name as the type . . . there are two versions of this test
to ensure that both field and method access on that type will work after I fix
the constructor weirdness.