Solve the type error I was encountering with the 'only' keyword
If a type was named in an 'only' list we used to not be able to access any of
its methods or fields, since they weren't also named in the list.  That seemed
rather unfortunate, so I decided to make inclusion of a type in an 'only' or
'except' list also apply to the type's fields and methods.  To do that, when
verifying that the symbols in the 'only' or 'except' list are accessible in
those scopes, we check if the symbol found is a type symbol and store the names
of its fields and methods if so.  Then, during function resolution (which also
handles field resolution) we check if the name provided to us is either in our
'except' or 'only' list or this expanded list of relevant names.

This change positively impacted the future related to functions returning a
type that wasn't accessible, but did not solve it.  It did solve the futures
related to the above problem, though.