Update initial comm counts, relecting a change due to #3061.
I recoded a bit of the flat locale model in #3061, resulting in 3 more
GETs and PUTs on each non-0 locale during locale model initialization.
Update comm counts to reflect this change.

Special thanks to @bradc for guessing correctly that my change was the
one that caused the nightly testing failures for this test.