Fix problem in virtualChildren
This patch is a bug fix. I was observing that virtualChildrenMap
sometimes contained functions that were not in the virtualMethodTable.
That meant that there was an inconsistency - some of these methods
would never be called and would be deleted (in some situations),
possibly causing the compiler to core dump.

This fixes a problem with
 test/users/ferguson/override_overload.chpl
which did not always cause the compiler to core dump, but in which the
virtualChildrenMap was always inconsistent (BaseWriter had
2 writeIt overrides for SubWriterTwo).

I fixed it in a straightforward but somewhat cumbersome manner because I
found it hard to understand and modify the existing virtual method table
generation code. My initial attempts at a simpler solution caused other
problems. I'll also note that this section of function resolution is
unnecessarily inefficient and that better data structures could almost
certainly give it much better performance.