Change Locale Models to implement writeThis only
It doesn't seem to make sense (at least at the moment) to read a locale. This
change makes it a compiler error to do so.

There are various compiler errors that I would expect in the read
path the way it is implemented now. So, until we care about (and test)
the read path, I'm disabling it for now.

For some reason the other changes in this PR trigger this error..