Add dead string literal elimination
When string-as-rec was committed (#2884) we saw a large increase in the
generated code size and thus compilation time. This is mostly because dead
string literals were being left in "ChapelStringLiterals". They were left there
because string literals are turned into Chapel records, and we don't currently
have any dead record elimination.

In order to resolve this, I initially tried to implement a general dead record
elimination. I ran into some issues with this so I opted to just do dead string
literal elimination for now. Dead string literal elimination is implemented as
follows:

String literals are all created with the `new_StringSymbol()` function. We now
mark the global variable that represents the Chapel string (the record) with a
flag. Then in dead code elimination we look for all these variables. If there
is only a single use of the string, we know that use is in the string
constructor, so we can safely DCE all the code used to construct the string
from the literal, as well as the literal itself. Note that we don't have to
worry about multiple Chapel strings using the same literal since we
hash/uniqify all string literals.

For jacobi, this takes us from 16,000 emitted statements down to 13,500 (was
12,000 before string-as-rec merge.)

Looking at just the number of emitted SLOC from `wc -l` (basically the number
we report, plus variable declarations) this takes us from 30,661 SLOC down to
25,500 (was 22,300 before string-as-rec merge.)

The remaining ~2,200 lines are coming from

| Module               | SLOC |
| -------------------- | ---- |
| ChapelDistribution   |  250 |
| ChapelStringLiterals |  300 |
| chpl__header         |   50 |
| Error                |  200 |
| IO                   |  700 |
| MemTracking          |  100 |
| String               |  900 |
| StringCasts          |  125 |
|                      |      |
| NewString (removed)  | -500 |

I'm not sure if there's too much we can do about them for now since most of the
code is for the new string implementation. At least a few hundred lines seems be
coming from dynamic dispatch branches being created instead of calls that index
into the chpl_vmtable. I still have to investigate that further as I'm not sure
what in the string-as-rec merge would have caused this.