Work around module ordering issue in NUMA configs
PR #2237 seems to have triggered in function resolution a recursion.  The
error occurs in numa configurations and hello3-datapar fails (among many
other tests).

The problem is that:
 * iterator might use a range
 * range might call halt
 * halt might call stringify
 * stringify might call c_calloc
 * c_calloc calls init_elts
 * init_elts might call the same iterator

I resolved the problem by making a param function,
isTupleOfPrimitiveTypes. The stringify function already included a branch
to work with primitive types only that did not call c_calloc. I made the
choice of which branch to take be a param value.

In order to get isTupleOfPrimitiveType to return a param, I needed to &&
reduce isPrimitive over a tuple. The only way I could think of to do that
was with recursive functions. An alternative would be to support
isTupleOfPrimitiveTypes with a primitive in the compiler.

Another alternative to this fix is to make c_calloc not call init_elts in
some cases (such as when the array is of primitive type).