Removing warning for this serialed reduction and adding notes for future
This test's reduction was getting serialized because it has bugs in
the reduction operator itself that make it unable to work in parallel
(and worse, unable to work correctly in parallel, I believe;
specifically, it wants to assert to the compiler that it is not
commutative, but doesn't have a good way to do so other than having
semantic problems in the combine() operator).  I wasn't sure how to fix
these issues and think it's an interesting case for us to be able to
support in a more principled way, so left some comments to this effect
since the test is passing.