Update futures that related to serialized reduction warnings
In most cases, this simply involved removing the warning from the future's
.good/.bad files.

In the case of kmeansonepassreduction-mystery, this test has been passing,
but was not caught because its .good file contained the warning which
hasn't been generated for this test for some time (and there was no .bad
file to alert us to it, presumably because the test predated the .bad
mechanism).

In the case of feature_locreduce, I entered a note as to why this
feature request isn't possible to consider without wholesale changes
to how our reduction framework is implemented.