Initial implementation of the local-on statement.
Like a regular on-statement, a local-on statement can be used to execute
a block of code on a sublocale. The difference is that the compiler will
not require wide pointers for symbols used across the local-on
statement. Local-on statements are compatible with the existing
coforall-on and begin-on optimizations.

If a user attempts to use a local-on to execute a block on a locale
with a different node ID, they will encounter a runtime error.

This commit adds the primitive PRIM_IS_WIDE_PTR. It takes a single
argument, a variable, and returns true if it is a wide pointer. It is
intended to be used to confirm that locality optimizations are firing.

This commit also enables local blocks if requireWideReferences() is
true. Previously, local blocks were not applied if a user was compiling
with CHPL_COMM=none and CHPL_LOCALE_MODEL=numa.