Remove c_string codepaths from the IO & Regexp module
The c_string codepaths are no longer needed and cause bugs when trying
to perform IO across locales. `string` is all that should be used for
textual IO.

Regexp was also using c_string in some places that would cause it to
cross locales, these have been removed as well.

The variables in ChapelEnv have been made into strings. This caused a
shuffle of the module load order in ChapelStandard.