Some clean ups
Stop using our homerolled Vec class when creating a UseExpr (and remove that
constructor option once I verified it wasn't necessary).  Also, make
scopeResolve's processImportExprs checks of the validity of the 'only' and
'except' keywords reuse code instead of duplicating it.  This meant I needed to
update the expected output for one of the tests I wrote, but that was fine.