Maintain the "task list" entirely inside the fifo tasking layer.
The tasking layer interface for managing structured parallel constructs,
embodied in the following 4 functions, has long seemed unnecessarily
complex:
  chpl_task_addToTaskList()
  chpl_task_processTaskList()
  chpl_task_executeTasksInList()
  chpl_task_freeTaskList()

It seems as though we should be able to make do with 3 or perhaps even
just 2 functions there, and also get rid of the opaque "task list" value
which is returned by addToTaskList() and then must be passed back into
the others.  Indeed, the Qthreads tasking layer really only uses the
first of these functions -- the other 3 are empty.  And, it ignores the
task list.  But the fifo tasking layer (which after all originated this
interface) has significant code in all 4 functions and uses the task
list extensively.

In this commit we streamline the task list management and make 2 of
these interface functions empty:
  chpl_task_processTaskList()
  chpl_task_freeTaskList()

Basically the task list ties together the the tasks created to run a
parallel construct, so that the thread running the task that originally
encounters that construct can help execute those tasks (since it has
nothing to do otherwise, while waiting for the construct to complete).
We used to do the following:
  addToTaskList:      allocate task list management data structures and
                      add each task to the list, one by one
  processTaskList:    take each task out of the list and add it to the
                      general pool of tasks to be run, except that as an
                      optimization, the encountering thread runs the
                      first task in the construct in a nested fashion
  executeTasksInList: the encountering thread runs any other unstarted
                      tasks in the construct in a nested fashion
  freeTaskList:       free the task list data structure

Now instead of allocating a whole set of list management data structures
we maintain the list inside the task descriptors.  In addition, we also
simplify operations on the list in the interface functions, as follows:
  addToTaskList:      add each task to the general pool of tasks to run
  processTaskList:    empty
  executeTasksInList: run any unstarted tasks in the construct in a
                      nested fashion
  freeTaskList:       empty

Note that currently this is resulting in some intermittent segfaults
and/or assertion failures that the queued_task_cnt isn't >= 0 when it
should be.  I believe I understand the problem there, and am working on
the solution.