Update syncPerfGraphs script to delete old files instead of using `rsync --del`
The script to sync the perf graphs to sourceforge used to call rsync with
`--del` to remove old graphs that were merged, removed, changed names, etc.
However, when we create the html/ directory for a machine we don't preserve
subdirectories so we might inadvertently delete subdirectories we don't want to.

For example we're syncing the `-sassertNoSlicing` results to chapcs/noSlice,
but that job finishes before the regular chapcs performance testing so when
chapcs is synced, it deletes the chapcs/noSlice data. Both jobs just leave
their graphs in chapcs/html instead of something like chapcs/noSlice/html/ so
there's no sane way to get rsync --del to work.