Add/update comments based on Ben's feedback
BenH pointed out that now that we have queries like .targetLocales()
and .localSubdomain(), we should use those here rather than the
poking around in the guts of Dom that I'm doing manually via
Dom._value....  This is an excellent point, but unfortunately those
queries are only supported on distributed arrays for now.  Ben and
and I are working to refresh our memory about whether that was
laziness or for a good reason.

This also reminded me that I should call out a bit more clearly
how I'm using a distributed global array to bootstrap a bunch
of distribution decisions that one would normally need to make
explicitly in an SPMD-style code like this because I'm lazy, so
I did that as well.