Add a makefile rule to always build chpldoc if env var is set
Add a top-level makefile rule to always build chpldoc if
CHPL_ALWAYS_BUILD_CHPLDOC is set. CHPL_ALWAYS_BUILD_CHPLDOC is a sibling to
CHPL_ALWAYS_BUILD_TEST_VENV, which was added with #2263

The rationale is that chpldoc is required for a full test run or you get an
error along the lines of "[Error matching man page(s) with compiler
environment]" and a bunch of tests are skipped.

I (and I believe a few other developers) never remember to build chpldoc so I
just mentally filter out that error. This is just for testing convenience.