Convert a "c for loop" in default rectangular to use a direct range iterator
#407 converted a loop over a range to a direct invocation of a "c for loop" in
DefaultRectangular. This was motivated by a performance regression that was
caused by the introduction of a range constructor (which was expensive in tight
loops.)

Using the "c for loop" directly was ugly, and meant that we didn't get any
overflow or legal stride checking. It also meant that the generalIterator()
wouldn't be invoked if a user compiled with `suseOptimizedRangeIterators=false`

The anonymous range iteration optimization (#1021) added some direct iterators
that take low, high, and stride as arguments and avoid range creation. There
was already one optimized for positively strided ranges, so this commit changes
the iteration over the "c for loop" to invoke that iterator.

For some reason it didn't occur to me at the time of that commit to make this
change, but some recent work on a range overflow bug (#2905) and some
improvements to the anonymous range iteration code (#3154) pointed me towards
this.